+++ src/main/kotlin/graphql/dataTransferObjects/InvalidMessageId.kt
@@ -3,6 +3,7 @@ package com.neelkamath.omniChatBackend.graphql.dataTransferObjects
 import com.neelkamath.omniChatBackend.graphql.routing.Placeholder
 
 object InvalidMessageId :
+    ReadMessageResult,
     CreateTextMessageResult,
     SetPollVoteResult,
     ForwardMessageResult,